Proyecto Personal/backend/requirements.txt
flask
mysql-connector-python
python-dotenv
flask-cors